tests: Improve dump_holes test
Change the volume layout to simulate the leases volume. The first slot
is always empty (was used by safelease in the past). Then we have some
reserved slots, and then volume leases.

The test shows that when we specify size, block size and alignment,
sanlock handles holes properly in the start, middle and end. This means
that we can always dump the leases volume from offset 0, so we don't
need to amend relative offsets in old sanlock versions.

Tested with sanlock 3.8.0 and 3.8.1 on Fedora 31.

Change-Id: I2711bb927fba6cd2648b4d6242044789cbc8e4d1
Signed-off-by: Nir Soffer <nsoffer@redhat.com>