net: Add used vlanned nic validation
Add nic usage validation, for not allowing vlanned
devices to share nics with bonds.

Adding this change rather than removing it at all was
agreed because this validation was already implemented in the
legacy switch setup. Adding this change to vdsm validation
rather than removing it from the legacy switch makes more
sense in the long run, where we support both legacy switch
and nmstate for now.

Even if nmstate allows this kind of scenario, VDSM will not
support it in order to reduce complexity and potential
configuration edge cases. The behavior is also in sync with
engine logic, where such a setup scenario is not supported.

With this fix the validation for used nics is performed in the
setupNetworks, regardless of the switch and backend.

Change-Id: I3a0d824a3e6338a52ed278ef981e6140248e9c1f
Signed-off-by: Bell Levin <blevin@redhat.com>