net, tests: move second networksetuphook test to pytest
Hook tests are one of the last net functional tests that has not been
moved to pytest yet. The second one is testing after_network_setup hook
by creating cookiefile and comparing it with the created network.

As it is last test that is being converted, the old test file is
removed as well.

Change-Id: I6adf4d6a53c6eca916df9c997c5c438fe1807de0
Signed-off-by: Andrej Cernek <acernek@redhat.com>