net tests: Remove old vlanned bonded rollback test
This test checks to see the bond and vlanned net are rolled back
to the original stage after the setupNetwork editing the net, fails,
and no bond or vlan is added.

This is already checked throughout the rollback_test.py in the net
functional tests, namely in test_edit_network_fails,
test_setup_two_networks_second_fails.

Testing no bond / vlan was added in test_remove_broken_network as well.

Change-Id: I11f7e81826dea5b9182938a673de203fa0231b24
Signed-off-by: Bell Levin <blevin@redhat.com>