net tests: Remove blockingdhcp test
This test is not relevant anymore since ifcfg is deprecated,
and nmstate succeeds in adding a network even if no dhcp server
in place.

It is a slow test, which was not run up until now. Setting up
a slow marking for pytest, and converting this test only for legacy
switch is not worth the effort and running time.

Change-Id: I3e9bd88fe7a84f1b967a8bc2431984d49c135492
Signed-off-by: Bell Levin <blevin@redhat.com>