tests: reload: Add option to disable udev
On RHEL 8.2.1 we see random failures in lvcreate and lvchange:

    Failed to udev_enumerate_scan_devices.
    Volume group "bz1837199-000000000000000000000-0006" not found.
    Cannot process volume group bz1837199-000000000000000000000-0006

David Teigland suggested[1] to try obtain_device_list_from_udev=0 to
mitigate this.

Add --no-udev option, setting obtain_device_list_from_udev=0 in lvm
--config. This is useful to test if this fixes the issue on RHEL 8.2,
and test the effect on RHEL 7.8.

This is the second time we disable obtain_device_list_from_udev. We did
the same in RHEL 6:

commit 23ce1d87fa98f35d83005e2958ff4065b78ff9d8
Author: Nir Soffer <nsoffer@redhat.com>
Date:   Mon Nov 4 22:08:45 2013 +0200

    lvm: Do not use udev cache for obtaining device list

We removed the setting in RHEL 7:

commit e820cc58d0f89480ed4ce701c492501539a0b7ba
Author: Fred Rolland <frolland@redhat.com>
Date:   Sun Oct 18 14:51:31 2015 +0300

    lvm: Use udev cache for obtaining device list

So either this was not fixed in RHEL 7, or it broke again in RHEL 8.2.
Having the option will make it easier to test if this is really fixed
next time.

Using udev should speed up LVM commands, but testing on CentOS 7.8 show
that it is slightly faster without udev. We perform the same work but
number of reloads is about 5% higher.

$ time ../reload.py --lv-count 50
...
2020-05-31 17:47:10,809 INFO    (reload/lv) Stats: reloads=567 errors=0 error_rate=0.00% avg_time=0.607 med_time=0.511 min_time=0.133 max_time=1.696
2020-05-31 17:47:10,810 INFO    (reload/pv) Stats: reloads=585 errors=0 error_rate=0.00% avg_time=0.588 med_time=0.488 min_time=0.135 max_time=1.708
2020-05-31 17:47:10,913 INFO    (reload/vg) Stats: reloads=518 errors=0 error_rate=0.00% avg_time=0.664 med_time=0.534 min_time=0.153 max_time=1.718

real 5m44.438s
user 2m7.326s
sys 2m7.907s

$ time ../reload.py --lv-count 50 --no-udev

2020-05-31 18:00:39,387 INFO    (reload/pv) Stats: reloads=616 errors=0 error_rate=0.00% avg_time=0.558 med_time=0.489 min_time=0.135 max_time=1.274
2020-05-31 18:00:39,561 INFO    (reload/lv) Stats: reloads=603 errors=0 error_rate=0.00% avg_time=0.570 med_time=0.494 min_time=0.132 max_time=1.438
2020-05-31 18:00:39,666 INFO    (reload/vg) Stats: reloads=544 errors=0 error_rate=0.00% avg_time=0.632 med_time=0.556 min_time=0.154 max_time=2.055

real 5m44.036s
user 2m0.997s
sys 1m52.350s

[1] https://bugzilla.redhat.com/1812801#c3

Change-Id: I2818edc030666bdd853a8892052507af759b0209
Bug-Url: https://bugzilla.redhat.com/1842053
Related-To: https://bugzilla.redhat.com/1837199
Signed-off-by: Nir Soffer <nsoffer@redhat.com>