tests: reload: Add option to disable --select
Using --select 'vg_name = xxx' or '--select 'pv_name = /dev/mapper/yyy'
LVM process metadata of all vgs, and then filter the results for
displaying. If processing metadata of unrelated vg fails it is being
updated by another LVM command, the reload may fail.

Add --no-select option so we can compare robustness and timing with and
without --select.

When reloading lvs with --no-select, we select only the lv, which has
the important property of not failing the command when the lv does not
exist, which is required for the current test.

Testing on CentOS 7.8 show that using --select is slower compared with
explicitly specifying pvs and vgs. We perform the same amount of work in
similar time, but number of reloads is about 40% higher.

$ time ../reload.py --lv-count 50 --no-udev

2020-05-31 18:00:39,387 INFO    (reload/pv) Stats: reloads=616 errors=0 error_rate=0.00% avg_time=0.558 med_time=0.489 min_time=0.135 max_time=1.274
2020-05-31 18:00:39,561 INFO    (reload/lv) Stats: reloads=603 errors=0 error_rate=0.00% avg_time=0.570 med_time=0.494 min_time=0.132 max_time=1.438
2020-05-31 18:00:39,666 INFO    (reload/vg) Stats: reloads=544 errors=0 error_rate=0.00% avg_time=0.632 med_time=0.556 min_time=0.154 max_time=2.055

real 5m44.036s
user 2m0.997s
sys 1m52.350s

$ time ../reload.py --lv-count 50 --no-udev --no-select

2020-05-31 18:08:39,412 INFO    (reload/pv) Stats: reloads=698 errors=0 error_rate=0.00% avg_time=0.489 med_time=0.420 min_time=0.132 max_time=1.564
2020-05-31 18:08:39,489 INFO    (reload/vg) Stats: reloads=869 errors=0 error_rate=0.00% avg_time=0.393 med_time=0.389 min_time=0.129 max_time=0.861
2020-05-31 18:08:39,520 INFO    (reload/lv) Stats: reloads=918 errors=0 error_rate=0.00% avg_time=0.372 med_time=0.375 min_time=0.128 max_time=0.808

real 5m41.847s
user 1m43.990s
sys 1m55.916s

With 50 lvs per vg we don't reproduce any failures. We need to test with
500 lvs to reproduce the errors.

Change-Id: I6d7a775b7ce0f39d57c62b2052074e38d48ba36a
Related-To: https://bugzilla.redhat.com/1837199
Signed-off-by: Nir Soffer <nsoffer@redhat.com>