backup.py: add 'parent_checkpoint_id' to BackupConfig
Currently, VDSM wrongly used from_checkpoint_id as the parent checkpoint
when starting a backup with a checkpoint.

This works only when backing up since the last backup, which is the normal
flow. But if last backup has failed after starting, the user need to use
the last successful backup checkpoint ID.

Users may also have other reasons to use an older checkpoint ID when
starting a backup, for example, if backup media was removed by mistake.
Another flow is when the user need to create a full VM backup when the
VM has a chain of existing checkpoints.

VDSM cannot distinguish between two VM backup possibilities -
  1. Starting an incremental VM backup.
  2. Starting a full VM backup when having a chain of defined checkpoints.

For that reason a new parent_checkpoint_id argument was added to
distinguish between the flows and set the parent_id of the created
checkpoint properly.

The following combination of 'parent_checkpoint_id' and 'from_checkpoint_id'
attributes can be used:

  1. Full VM backup without a previous checkpoints chain -
     parent_checkpoint_id=None, from_checkpoint_id=None

  2. Full VM backup with a previous checkpoints chain -
     parent_checkpoint_id=111-222, from_checkpoint_id=None

  3. Incremental VM backup -
     parent_checkpoint_id=111-222, from_checkpoint_id=333-444

  4. Incremental VM backup without specifying parent_checkpoint_id -
     Will raise a backupError

Starting a backup will fail also if:

  1. parent_checkpoint_id provided but the defined checkpoints list
     on libvirt is empty.
  2. parent_checkpoint_id doesn't mach the defined checkpoints list
     leaf on libvirt.

Change-Id: Id0416eee1e5eb97445f27ab27a2b287a43bb9ca8
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>