tests: reload: Add option to control lvm verbosity
When handling lvm command failures, LVM developers typically ask to run
the command with -vvvv. Make it easy to provide this by adding --verbose
option.

When using verbose mode, LVM generates lot of output (1.5-4 MiB in this
script), and logging it is not helpful. Errors are dumped to files named
"command-error-nnnn.txt" with all the info about the error. This
makes it very easy to file useful bugs.

When using --debug mode, only the first 200 characters of stdout and
stderr are logged when logging command completion.

Since verbose creates huge output and we care only about the reloads
commands, create 2 separate lvm runners - one for the workers and one
for the reloaders. The reloader instance is using args.use_udev and
args.verbose, while the workers use the defaults.

Change-Id: Ifb0e457cf3478fcfc1f2b8cc1cc7c2b5e0beb2de
Related-To: https://bugzilla.redhat.com/1837199
Signed-off-by: Nir Soffer <nsoffer@redhat.com>