gluster: Added check and change for ipv6 hostnames in gluster vol list
The following gluster command fails in vdsm when trying with ipv6 hostnames

[root@ ]# gluster --remote-host=host1-storage.lab.eng.blr.redhat.com volume info engine
Connection failed. Please check if gluster daemon is operational.

It has been observed that adding '--inet6' to ipv6 hostname works.

[root@newhost ~]# gluster --remote-host=myhost.lab.eng.blr.redhat.com --inet6 volume list
testrep

This patch checks if a given hostname is ipv6 enabled and adds '--inet6'
to the gluster command.

Added GlusterFQDNToIpResolveException to the list of known exceptions
Change-Id: I8f8bb26e8515b739d3086ecfec8516d84e9455cf
Bug-Url: https://bugzilla.redhat.com/1841076
Signed-off-by: Kaustav Majumder <kmajumde@redhat.com>