lvm: Remove --select usage from reload lvs
Revert back to using "lvs [vg1/lv1, vg1/lv2, ...]" for reloading lvm cache.
Using --select option for lvm command requires a scan of all VGs
metadata before it can process only the relevant devices for the selected
entities. In vdsm 4.3 with lvm2-2.02 this induces the risk for transient
corruption in VG metadata for reload commands, for vdsm 4.4 with lvm2-2.03
this is not reproducible but there is a performance overhead for using
lvm --select commands. For sake of stability in released and undergoing
release versions we revert back to the old way.

This implies that reload with "lvs vg1/lv1 vg1/lv2" where lv1 is missing
from LVM will fail the entire command and return Unreadable LV entries
in cache for both lv1 and lv2.

Change-Id: I7ebf07135bb8b374688d9ed38f3a0f55f8b0587d
Bug-Url: https://bugzilla.redhat.com/1837199
Signed-off-by: Amit Bawer <abawer@redhat.com>