lvm: Remove wants_output check from cmd method
Using --select for lvm reload commands introduced a problem as
we depend on command failure to refresh a stale filter and --select
usage would return rc=0 even if querying some items has failed.
This used to be fixed by refreshing the devices filter for lvm command
also when a command expected output, but no output was received.

Since we revert from using --select for lvm reloads, this parameter
handling is now removed.

Change-Id: I9082d0e96f684fbd2f0345f76d2d131294516df5
Bug-Url: https://bugzilla.redhat.com/1852344
Signed-off-by: Amit Bawer <abawer@redhat.com>