blocksd: remove SD mappings upon SD invalidation
Currently, when removing stale SD from a cache, we try to remove also
all SD mappings. This is wrong as evicting the cache shouldn't impact
running system in any way. There can be some temporal glitch by which
the VG is not reloaded, considered as stale and all it's mappings are
removed.

Remove VG mapping cleanup from LVMCache._reloadvgs() and log only
a warning when removing stale VG (as we do for PV or LV).

Change-Id: I930f82867a97d2cb3ebe9e72f7ebb4404bb6e64a
Bug-Url: https://bugzilla.redhat.com/1849558
Signed-off-by: Vojtech Juranek <vjuranek@redhat.com>