lvm: Revert change for returned vgs entries upon reload error
This patch reverts the change:

commit c1db4021d02475549cc088b7dd83d1a67289f0d2
Author: Nir Soffer <nsoffer@redhat.com>
Date:   Sun Mar 15 15:05:31 2020 +0200

    lvm: Fix _reloadvgs() return value on errors

The reverted change made sure that vgs reload would return the Unreadable vgs
as well as the updated vgs to correlate with the updated LVM cache contents for
the queried VGs. Removing --select and its soft error handling reveals
a bug where Unreadable entries within the VG cache would never be cleared
from it since they were added to the updatedVGs dict which is considered to hold
the most up-to-date status of all queried VGs in the vgs reload flow.

Change-Id: Id948e843f4d01207101b692842defe84070d6f33
Bug-Url: https://bugzilla.redhat.com/1837199
Signed-off-by: Amit Bawer <abawer@redhat.com>