lvm: Remove --select usage from reload vgs
Revert back to using "vgs [vg1, vg2, ...]" for reloading lvm cache.
Using --select option for lvm command requires a scan of all VGs
metadata before it can process only the relevant devices for the selected
entities. In vdsm 4.3 with lvm2-2.02 this induces the risk for transient
corruption in VG metadata for reload commands, for vdsm 4.4 with lvm2-2.03
this is not reproducible but there is a performance overhead for using
lvm --select commands. For sake of stability in released and undergoing
release versions we revert back to the old way.

This implies that reload with "vgs vg1 vg2" where vg1 is missing from
LVM will cause the entire command to fail, marking vg1 as Unreadable,
and parsing vg2 information since the code for VGs reload does not break
upon failure and parses the valid VGs it has found.

Change-Id: I2eb015d200da3e0962cbddec59777d7f02340dba
Bug-Url: https://bugzilla.redhat.com/1852344
Signed-off-by: Amit Bawer <abawer@redhat.com>