http: Close connection after errors
The http server uses keep alive connections but this does not work well
when the internal task fails while reading the body from the client or
sending the body to the client.

The best way to handle errors in this case is to close the connection
right after the error. This allows the client to detect the error early
instead of timing out after a long delay.

Change-Id: Icf985349916f5e7c72146a7efe366e68eb88d8ca
Signed-off-by: Nir Soffer <nsoffer@redhat.com>