lvm: Revert change for returned vgs entries upon reload error
This patch reverts the change:

commit 8f66881359cce06004a39bd9d85cecd996c7b5bb
Author: Nir Soffer <nsoffer@redhat.com>
Date:   Sun Mar 15 15:05:31 2020 +0200

    lvm: Fix _reloadvgs() return value on errors

The reverted change made sure that vgs reload would return the Unreadable vgs
as well as the updated vgs to correlate with the updated LVM cache contents for
the queried VGs. Removing --select and its soft error handling reveals
a bug where Unreadable entries within the VG cache would never be cleared
from it since they were added to the updatedVGs dict which is considered to hold
the most up-to-date status of all queried VGs in the vgs reload flow.

Change-Id: Ie1c7a6a3cac03c1e2cf5833cbaa3d98b35c3bd3a
Bug-Url: https://bugzilla.redhat.com/1852344
Signed-off-by: Amit Bawer <abawer@redhat.com>