net: Add vdsm common patching for unit tests
The network unit test should not compile vdsm, since we should
not rely on the developer's system, and have a universal
environment. The common files are required by almost
all network services and therfore needs to be patched.

An el8 pregenerated static version was added to be able to
run the unit tests without compiling vdsm (this version
is the bare minimum of variables to run to be able to pass
the imports of the network conftest)

Patching the files during the run test script is preffered,
rather than patching through pytest (in conftest.py).
Constants is imported as early as in the conftest itself,
which leads to an exception before any test is able to run.

Change-Id: I4ffa3f6f2f2ea67b44ae2bfbb9c8f7d0cdc84883
Signed-off-by: Bell Levin <blevin@redhat.com>