devices: remove backingstore in restore
Previously, when creating snapshot metadata the saved data didn't
contain backingStore element in the disks. This is due to the changes
made to the VIR_DOMAIN_XML_MIGRATABLE flag output in libvirt >= 6, which
is used to produce memory dump configuration file, and since the backing
chain is now kept stable, it is now part of the output when the flag is
used. When we create multiple snapshots in a single VM run, the backing
store was saved and upon loading it created a confusion.

This patch will remove the backingStore element in disks on restore
memory snapshot flow to prevent backing chain confusion.

Change-Id: I4891f03dbdd7ea92661da25bd820ff9c06243b6a
Bug-Url: https://bugzilla.redhat.com/1840609
Bug-Url: https://bugzilla.redhat.com/1842894
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>