tool: Add return codes to config_lvm_filter when cannot configure
Provide indication to caller about resulting execution status
of lvm filter configuration.

By default, vdsm-tool config-lvm-filter would return 0 on successful
completion. 1 is returned when exception occurs during execution and
2 for error in parsing wrong arguments.

This patch adds return code 3 to indicate the tool has detected that
lvm filter has to be reconfigured but refrained from doing so since
there is already another lvm filter configured on the host.

Return code 4 is added to indicate that the tool has detected that
lvm filter has to be reconfigured but user aborted the operation by
selecting "NO" in confirmation input (when executed without the
--assume-yes|-y option).

Change-Id: Ie1024c3563fbed5855cd5b49b5a3af5e437d3cf5
Bug-Url: https://bugzilla.redhat.com/1522926
Signed-off-by: Amit Bawer <abawer@redhat.com>