net, nmstate: Introduce basic OvS networking
Add module that will allow configuration of ovs
through nmstate. The initial part contains only the
bare minimum to configure network that is using ovs.

The structure of the ovs network is generally simpler
than linux bridge network. On the diagram [0] you can
see the basic structure.

The SB represents southbound interface that is usually
physical nic or bond. For every SB exists only single
bridge that has SB as one of its ports.

Any network defined from oVirt point of view is defined
in form of NB (northbound). NB in ovs is represented
as a single ovs internal interface connected via port
to the bridge. This simplifies the VLAN management
because the filtering is specified per port.

The former implies that in ovs there is no form
of bridgeless network as every network is connected
to a given bridge.

Another slight caveat with this approach is that bridge
parameters, most importantly STP, is hard to enforce
on the bridge as the connected networks might oppose
whether it should be used or not.

[0]
+------+  +------+  +------+
|      |  |      |  |      |
|  NB  |  |  NB  |  |  NB  |
|      |  |      |  |      |
+--+---+  +--+---+  +---+--+
   |         |          |
   |    +----+-----+    |
   |    |          |    |
   +----+  OvS br  +----+
        |          |
        +----+-----+
             |
             |
        +----+-----+
        |          |
        |    SB    |
        |          |
        +----------+

Change-Id: Ide74ae772018b25dd384fa4cbdb1e809d453183b
Bug-Url: https://bugzilla.redhat.com/1809102
Signed-off-by: Ales Musil <amusil@redhat.com>