virt: Prevent migration monitoring race and traceback
Migration monitoring thread checks for end of migration threading
event and then retrieves the migration job stats.  There is a timing
issue there: If the migration finishes and the source VM is destroyed
just before the stats are retrieved, the job stats retrieval will
raise an exception.  The exception is harmless, it just causes the
monitoring thread to finish, which should happen to it anyway.  But it
logs the error and traceback in Vdsm log.

This patch prevents the error and traceback logging by handling the
exception.

Change-Id: Ib5ec45f4b362e855678bbce483b8cf34d82c590e
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>