net, nmstate, test: Reuse RunningConfig and current_state mocks
Move current_state_mock to conftest instead of
definition in every file.

Move rconfig_mock to conftest instead of definition
for every function that needs it. As bonus it can't
happen that polluted env could brake the test.

current_state_mock is used by every test for nmstate.
Move it to conftest instead of definition in every file.

Change-Id: I6faf32d3b7cad9ef86a106b8a3790794e09cce32
Signed-off-by: Ales Musil <amusil@redhat.com>