gluster: Modified check for ipv4 fqdn instead of ipv6
A change was added in patch: https://gerrit.ovirt.org/109360
where ipv6 fqdn is resolved with
socket.getaddrinfo(address, None, family=socket.AF_INET6) and '--inet6'
However it has been noticed that with hosts having dual nics(both ipv4)
ipv6 was resolved via the fqdn.
This patch should fix the above by checking only ipv4. If it is not
resolved ipv6 is regarded and corresponding `--inet6` is added to the
gluster command.

Change-Id: Idf185f900aa0305c8251bf9aeab26c60959fe7ae
Bug-Url: https://bugzilla.redhat.com/1841076
Signed-off-by: Kaustav Majumder <kmajumde@redhat.com>