qga: add partition stripping for libvirt backend
This was originally added to poller in 66dd1d84 but was missed later
when adding the libvirt backend. What was returned from qemu-ga was
inconsistent with what we did return with ovirt-guest agent. Instead of
mapping disk devices to serial numbers we in fact mapped a partition.
E.g. /dev/sda1 instead of /dev/sda.

The current behavior did not make much sense anyway. Suppose there were
two partitions with filesystem on the disk /dev/sda1 and /dev/sda2 then
either of those could get associated with the disk serial number.

Change-Id: I6034c75a74bff7082716d42a0c48631aee4ac632
Bug-Url: https://bugzilla.redhat.com/1793290
Signed-off-by: Tomáš Golembiovský <tgolembi@redhat.com>