lvm: Avoid iterating pvs dict without locking
removeVG performs tuple comprehension outside of LVMCache
over pvs dict to find PVs of the VG to be invalidated
after the VG was removed from lvm.

This can be unsafe since pvs dict is common to all VGs in cache
and if another VG operation modifies the pvs dict during
tuple comprehension we can get a RuntimeError.

Remove the tuple comprehension and use the lock protected
_invalidatevgpvs method for invalidating the removed VG's pvs.

Change-Id: I09b1f3cafadbbfadddd1ce93a01ef310f874d8c3
Bug-Url: https://bugzilla.redhat.com/1856065
Signed-off-by: Amit Bawer <abawer@redhat.com>