net, nmstate: Expose OvS through generate_state
By exposing the OvS switch it is now possible to
configure networking with the combination nmstate
and ovs. However this is by no means complete.
The exposure is currently used only for unified
unit test API.

It is highly advised to actually NOT use ovs over
nmstate in the current state.

Change-Id: I30f468a70133411d4410ed3fe8006e0ae18fa328
Bug-Url: https://bugzilla.redhat.com/1809102
Signed-off-by: Ales Musil <amusil@redhat.com>