machinetype: Add spec_ctrl feature also to new CPUs without -IBRS
New CPU models are available that have spec_ctrl feature by default.
With more of those CPUs coming, we need some way mechanism to detect
the feature without checking -IBRS suffix in the model name, which is
not present in new models.

Let's use heuristic where we look for -IBRS suffix in any of the
compatible CPU models.

Change-Id: I4e19dc600055da8efc323e524206abd707ad3391
Bug-Url: https://bugzilla.redhat.com/1854922
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>