monitor: Add monitors dictionary lock
Iterating over the monitors dict while other methods can
remove or add monitors to it can result in a RuntimeError
for python3.

The resolution in this case is to add a class lock for the
monitors dict to prevent its modifications from one call
while it is iterated from another.

For longer operation like shutdown we add shutdown flag
to indicate for stop or start monitoring methods that it is
taking place and in such case they would abort by raising
ShuttingDownError. That would prevent any monitors dictionary
modifications while shutdown is in progress outside the lock,
allowing any other read-only dict operations to take place.

Change-Id: I6e89a24a03d0623ae69e300989dc60344c53eb6a
Signed-off-by: Amit Bawer <abawer@redhat.com>