pytest: Unify logging configuration
Currently logs from pytest runs were configured
differently in each module which created inconsistency.

Unify it to have common pattern that includes
information that might be helpful for finding the issue.
With that we can remove the individual settings.

Only place that remains is the testrunner.py which still
uses nose. Update its format to match the global one.

Change-Id: I1f5b71cd1d188313e30e9ef3e8f801f418dbf278
Signed-off-by: Ales Musil <amusil@redhat.com>