fc-scan: Limit number of workers on smaller machines
Looking at fc-scan on an older system with 8 cores, is seems that using
64 workers causes unwanted load for no benefit. It is also not needed
since smaller systems are likely to have less running VMs and less LUNs.
Limit the number of workers to 1 worker per CPU, up to 64 workers.

Tested on a system with 64 paths to storage. Without this change fc-scan
will try to scan all devices at once. With this change, 8 devices at the
same time.

--------------------------------------
workers   total  time     rescan  time
          real    sys     avg      max
--------------------------------------
  8       629     686     0.26   18.89
 64       624     815     0.94   37.09

Data was collected using this script:

$ cat fc-scan-test.sh
for i in $(seq 100); do
    /usr/libexec/vdsm/fc-scan -v
    udevadm settle --timeout 10
    sleep 1
done

Change-Id: Iebf9dc8f71972e0999166dd95b74174ea6b6fcb1
Related-To: https://bugzilla.redhat.com/1851893
Signed-off-by: Nir Soffer <nsoffer@redhat.com>