dump-volume-chains: use storage.dump() api
This tool can make good use of the new dumpStorageDomain() API.
It will speed up data collection and lower the system overhead
when doing it by using a single API instead of nested loops
with image and volume related APIs.

It will help not hitting the sos default timeout of 300s
when running such commands, which can cause incomplete
sosreports.

Change-Id: I8082c52acd410d0f072c949869a1210cf7bad162
Bug-Url: https://bugzilla.redhat.com/1839444
Signed-off-by: Germano Veit Michel <germano@redhat.com>