API.py: Add new addBitmaps flag to Volume.create() verb
Now, qemu-img supports the 'bitmaps' sub-command to perform operations
on bitmaps that were created during a backup operation.

A new flag added to Volume.create() verb, this flag will instruct
the host, in case of snapshot creation, to add all the bitmaps from
the source volume to the newly created volume with the same bitmap names.
Currently, in this case, the bitmaps will not be added to the newly
created volume and when a backup operation will take place it will be
failed.

This patch only adds the API for adding the bitmaps when a snapshot is
added to a volume, the following patches will implement adding the
bitmaps of the for file and block volumes.

Change-Id: I6d3e080b28efd9474f9441777eb33085c558aeb3
Bug-Url: https://bugzilla.redhat.com/1861667
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>