volume.py: add volume bitmaps after snapshot created
If a volume contains QEMU bitmaps and a snapshot
created for that volume, the bitmaps will not be copied to
the new 'active' volume, which means that the next backup
operation on that volume will fail.

Calling 'qemu-img bitmaps' sub-command to add all the bitmaps
from the source volume to the destination volume.

New bitmaps.py helper introduced to include all bitmaps related
logic.

Change-Id: Id0c18e3a7f8bf5c4192e8ff04c8ad2ee64e47870
Bug-Url: https://bugzilla.redhat.com/1861667
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>