tools: Start and reconfigure multipathd
In the following host setup scenario:

- host is booting from a multipath device.
- host is using the default multipath configuration option:
 "use_friendly_names = yes".
- host is not running multipathd service by default.

the root file system is on device /dev/mapper/mpath{X}. As part of
Vdsm configuration, we disable user friendly names, and reload
multipath configuration. This renames /dev/mapper/mpath{X} to
/dev/mapper/{WWN}.

Right after we reload multipath we create lvm filter based on the
device used for the root filesystem. If multipathd was not running
when we reloaded the service, the lvm filter will be created using
the wrong device name (/dev/mapper/mpath{X}) and the next boot will
fail.

The purpose of this change is to ensure that existing multipath
devices using "friendly names" are renamed before lvm fiter is set
on the host to prevent the missing rootfs device issue.

Change-Id: I78051831628ce62fcd96a75db149846819b3366d
Bug-Url: https://bugzilla.redhat.com/1859876
Signed-off-by: Amit Bawer <abawer@redhat.com>