tool: multipath: Improve documentation
The existing documentation was focused on single use case. Replace with
more general text explaining all possible use cases and how the current
code deals with all of them.

This area of the code was broken many times in the past because it was
poorly documented and developers did not understood the different use
cases. I hope that with the new documentation this will not happen
again.

Change-Id: Ic34767fcab0ba31dc2afeba9aa1b897fc72cdb9a
Signed-off-by: Nir Soffer <nsoffer@redhat.com>