net: Fix validation for used vlan
We do not allow unmanaged vlans to have same base interface
with bond. The problem was that current code did not care
if the vlan was unmanaged. That resulted in "Interface in use"
validation error for vlans mamanged by vdsm.

Include only the unmanaged vlans in the validation. Since
it is not possible to take over vlan and define a bond over
it's base interface in one operation, the check can be done
over current network state and not desired.

Change-Id: I174b2c44ce168548dc66b58dfb7d049c3d567a52
Bug-Url: https://bugzilla.redhat.com/1870148
Signed-off-by: Ales Musil <amusil@redhat.com>