qga: the default setting for context should be blocking
The `default` timeout value means 5 seconds for `guest-sync` and 5
seconds for the requested command. This is not the same as the behavior
before introducing the timeout setting API in libvirt. In fact the
original behavior corresponds to the `blocking` setting which is not
100% blocking but it still has 5 seconds timeout on the `guest-sync`
command and blocks only for the duration of the requested command.
The `blocking` setting matches exactly the behavior prior introducing
the libvirt timeout API. Without it some long running tasks (e.g.
filesystem freeze) can timeout unexpectedly.

Change-Id: I6697b6dfe0061526d5d811ebc2f9304e46903382
Signed-off-by: Tomáš Golembiovský <tgolembi@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1870500