qga: release lock after timeout reset
To avoid accidental races when setting the timeout we should first reset
the default value and then release the lock. Otherwise we could possibly
reset timeout that was already set in different thread.

To make sure the lock is always released, even in case ov libvirt error,
another try-finally block is added.

Change-Id: I1ad5b896c3ee88bd1fce4cef65c3a4322eaf4de3
Signed-off-by: Tomáš Golembiovský <tgolembi@redhat.com>