virt: Don't send device hash for non-libvirt domain XML
When a VM is in the process of starting, Engine may call
getAllVmStats, find a device hash there for the starting VM and since
it has no previous device hash it asks for the VM domain XML by
calling dumpxmls.  However, the VM may not have its domain XML
obtained from libvirt and then dumpxmls responds with the domain XML
initially obtained from Engine.  The domain XML doesn't contain added
information from libvirt for newly created VMs yet, but Engine expects
it and may complain with messages such as:

  ERROR ... managed non pluggable device was removed unexpectedly from libvirt: ...

or

  DEBUG ... managed pluggable device was unplugged : ...

It should normally get fixed in a while, by a later dumpxmls call,
unless the VM fails to finish its start, in which case the devices
would remain unplugged.

This patch distinguishes between domain XMLs obtained from Engine and
libvirt and doesn't report device hashes for the former.  This is an
API change but Engine is happy with it and stops unplugging the
devices and logging the errors.

We consider _srcDomXML's also initial, because they are retrieved from
libvirt as migratable and needn't be valid at all from some points of
view.

Change-Id: Ib8bd8be88725d2342e957e87119ecf8151092f5d
Bug-Url: https://bugzilla.redhat.com/1870108
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>