static: Support all-in-one setup
All-in-one setup, when both engine and vdsm are running on the same host
was deprecated in 3.6 and became unsupported in 4.0. However users are
still using this setup for some reasons.

With this setup we have both engine and vdsm configuration in:

    /etc/ovirt-imageio/conf.d/
        50-engine.conf
        50-vdsm.conf

Since 50-vdsm.conf is sorted after 50-engine.conf, vdsm configuration
overrides engine configuration. However engine set:

[local]
enable = false

[control]
transport = tcp

When vdsm set nothing since this the defaults are good enough. So engine
configuration breaks the control socket, and vdsm fails to add tickets
with:

    Image daemon is unsupported

To fix this, vdsm configuration was renamed to 60-vdsm.conf, to make it
clear that both engine and vdsm configuration can be on the same host,
and the missing configuration was added to vdsm configuration.

Change-Id: Iadbed463a2adba9f4fd74d57438ac5f06d6ffb84
Bug-Url: https://bugzilla.redhat.com/1871348
Signed-off-by: Nir Soffer <nsoffer@redhat.com>