monitor: Use a list copy of the monitor values for stopping
The stopMonitors method loops over the monitors dict and removes
items from it once the corresponding monitor thread has stopped.

This leads to RuntimeError since we pass a view to the monitors
dict itself which the method later iterates over and modifies its size.

 File "/usr/lib/python3.6/site-packages/vdsm/storage/hsm.py", line 3429, in prepareForShutdown
    self.domainMonitor.shutdown()
  File "/usr/lib/python3.6/site-packages/vdsm/storage/monitor.py", line 254, in shutdown
    self._stopMonitors(self._monitors.values(), shutdown=True)
  File "/usr/lib/python3.6/site-packages/vdsm/storage/monitor.py", line 273, in _stopMonitors
    for monitor in monitors:
RuntimeError: dictionary changed size during iteration

This regression was introduced as part of monitor fixes patch:

 commit cecc2d1f9d77f80843ca8977936d3479f25bf609
 Author: Amit Bawer <abawer@redhat.com>
 Date:   Tue Jul 14 11:08:00 2020 +0300

     monitor: Add monitors dictionary lock

Change-Id: I4ce31f06f2c7470725d7281f4f8bb5b21721bb4c
Signed-off-by: Amit Bawer <abawer@redhat.com>