vm: Use state indication for live merge cleanup thread
This change allows to follow the cleanup thread by observing its states:

- TRYING: Starting state for a fresh thread, trying to pivot.
- RETRY: Recoverable failure during cleanp indicating to the caller
         that the thread run should be retried.
- DONE: Successful completion state for the cleanup thread.

In case the cleanup run has terminated with a recoverable error
the corrsponding log warning is printed:

WARNING (merge/job-name) Pivot failed (job: job-id): Block copy job job-id is not ready for commit, retrying later

And the thread will transition into RETRY state.

In a followup patch an ABORT state will be added to inidicate that
an unrecoverable error was met during pivot operation and merge cleanup
should not be reattempted.

Change-Id: Idbf8b28ea53f135dbee1d4669e86b7c3e0c76010
Bug-Url: https://bugzilla.redhat.com/1857347
Signed-off-by: Amit Bawer <abawer@redhat.com>