vm: Stop live merge cleanup attempts for unrecoverable pivot error
If pivot attempt has failed due to an unrecoverable libvirt error
the cleanup thread would still be re-invoked on the next
block jobs query cycle. There is no purpose for such retries in
case of a libvirt error which is not for an ongoing block copy.

This patch adds exception handling for BlockJobUnrecoverableError
in case libvirt pivot operation has raised a non-block-copy error,
this indicates that the cleanup thread will not be re-invoked
for the failing block commit job in a futile attempt to recover it
later as its state was set to ABORT.

In case cleanup has failed with an unrecoverable error, the following
error is printed to log:

ERROR (merge/job-name) Pivot failed (job: job-id): Block job job-id failed with libvirt error: error, aborting due to an unrecoverable error

Change-Id: Iefe7eef68a52bb50c90c6533682ace9964a4f7b8
Bug-Url: https://bugzilla.redhat.com/1857347
Signed-off-by: Amit Bawer <abawer@redhat.com>