virt: Return back the check for host shutdown
When a VM stops running due to shutdown, libvirt tells us whether the
reason was shutdown from within the guest or from the host.  In commit
21d8c83d700b99827784585fb91087cd9fe319f3, we replaced some of our
guesswork with the information from libvirt.

However that information is also not completely reliable.  When a host
is shut down by an administrator, e.g. by running `poweroff',
libvirt-guests service gets invoked and attempts to shut down the
running VMs gracefully, using `virsh shutdown' command.  This is a
good thing to do, but then it looks as a user initiated shutdown to
QEMU and libvirt (confirmed in
https://www.redhat.com/archives/libvirt-users/2020-August/msg00095.html).
If the stopped VM is a highly available VM then
it's not started on another host due to the wrong shutdown reason
reported to Engine.

Let's fix that by returning back one of our former checks for host
shutdown.  The aforementioned commit tried to fix a similar issue
where a host shutdown hasn't been detected, under different
circumstances, but it created the problem described above.  Adding the
additional check fixes that.  We also add the check when receiving an
unexpected shutdown detail, it doesn't harm there and can help
identify the right shutdown reason.

Change-Id: I158d4ad019ceead24a4908d8c8830c3b635e3dca
Bug-Url: https://bugzilla.redhat.com/1800966
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>