lvmfilter: replace unstable names in filter with stable ones
In the past we use unstable device names in the filter (e.g. /dev/sda1).
Such name can change during reboots and filter can become invalid.

Add new check into lvm filter if we can replace existing filter with
unstable names with one using stable names. The replacement is suggested
only in case we can replace all items and the filter wasn't modified by
the user. E.g. if user added more items (no matter if with stable or
unstable names), we won't do any replacements and suggest to fix the
filter manually. Also, as we want to replace names with ones containing
UUIDs, we cannot replace any regular expressions with wild cars. The
only exception is ignoring all devices ("r|.*|"), which is kept as is in
case of replacement.

If we detect, that filter can be replaced by new one with stable names,
CONFIGURE advice with new filter is returned, otherwise RECOMMEND with
new filter is returned.

Change-Id: Iea32ce0b9c1ce3a521c69aa850709a6590d90120
Bug-Url: https://bugzilla.redhat.com/1837864
Signed-off-by: Vojtech Juranek <vjuranek@redhat.com>