lvm: Disable LVs caching
Recent failures in OST are caused by enabling LVs caching in:

commit 58e3efe5bd1f04f51cbcb5e43ad09634818c1569
Author: Amit Bawer <abawer@redhat.com>
Date:   Mon Mar 16 12:56:31 2020 +0200

    lvm: Turn stalelv boolean indication into freshlv set

What we see in OST:

host-0 (SPM):

2020-09-04 18:24:43,925: Starting create LV (create LV).
2020-09-04 18:24:46,043: Completing create LV (change tags).

host-1:

2020-09-04 18:24:58,543: Prepare image fails, volume does not exist.

Preparing an image does:

    HSM.prepareImage()
        BlockStorageDomain.getAllVolumes()
            BlockStorageDomain.getAllVolumesImages()
                blockSD.getAllVolumes()
                    blockSD._getVolsTree()
                        blockSD._iter_volumes()
                            lvm.getLV()
                                LVMCache.getLv()
                                    LVMCache._lvs_needs_reload()

On the SPM, creating a LV invalidate the cache, so _lvs_needs_reload()
will return True. On other hosts, if LVs were reloaded recently, the
cache is considered fresh, and _lvs_needs_reload() will return False.
Then we use the cached LVs which do not include the new LV created on
the SPM.

Add cache_lvs flags to LVMCAche, set to False by default. Use the lvs
cache only when the flag is True. This is currently used only in the
tests, verifying the logic for using the cache.

We can use the new cache_lvs on the SPM to avoid unneeded reloads in the
future, or when we have a way to detect changes in the VG from other
hosts.

Change-Id: I9f64baf9f50434f2a7d4b3ffda80a604ba207526
Bug-Url: https://bugzilla.redhat.com/1876230
Signed-off-by: Nir Soffer <nsoffer@redhat.com>