API.py: Add new merge_bitmaps flag to SDM merge() verb
Now, qemu-img supports the 'bitmaps' sub-command to perform
operations on bitmaps that were created during a backup operation.

A new flag added to the SDM merge() verb, this flag will instruct
the host if all the bitmaps from the top volume should be merged
to the base volume.

This patch only adds the API for merging the bitmaps when a
snapshot is removed when the VM is down.
The following patches will implement merging the bitmaps for
file and block volumes.

Change-Id: I6ce4f85ec81ef157c25a0810300efd9a4db19ed0
Bug-Url: https://bugzilla.redhat.com/1861673
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>