sdm/api/merge.py: merge bitmaps in merge job
If a snapshot contains QEMU bitmaps and it's been removed,
the bitmaps will not be merged to the base volume, which means
that the next backup operation on the base volume will fail.

This patch adds the logic to merge bitmaps from the removed top
volume to the base volume during SDM merge, bitmaps will be merged
only if the base volume is in 'qcow2' format and the engine
requests to merge the bitmaps (merge_bitmaps=True).

Change-Id: I3686002080f210cc787e1e5670f7bce88c82f6e9
Bug-Url: https://bugzilla.redhat.com/1861673
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>