qga: remove dead appsList code
Originally we used to mimic the behavior of oVirt Guest Agent and report
qemu-ga as 'QEMU guest agent' in apps on Windows. During transition to
libvirt API this code turned out to be dead and we started reporting
qemu-ga as 'qemu-guest-agent-<version>' just like we do on Linux.
Instead trying to fix the code for Windows it is better to drop the dead
code altogether and keep reporting what we do now. Engine needs the
format with qemu-ga version for tracking of tools update process.

Bogus constant _GUEST_OS_LINUX is also removed because it is misleading.
There is no single value for linux OSes as it varies between
distributions (e.g. 'rhel', 'fedora', etc.).

Change-Id: I3923027601921fbd2d0bf4e675cb813d47c00a27
Signed-off-by: Tomáš Golembiovský <tgolembi@redhat.com>