net, tests: Use blocking dhcp as indicator for sync dhcp
If blocking dhcp is specified we should wait for the dhcp
response. In some tests we do not actually need to wait
for the result because dhcp server is not running.

Use the blocking dhcp and skip the wait in tests that do
not need it. This will cut running time of functional tests.

Change-Id: I27a89d2615427a77aee0415f82266931be5a124a
Signed-off-by: Ales Musil <amusil@redhat.com>