vmdevices: forward compatibility for lun guestName
The engine expects imageID element when it sees guestName element.
Without populating it, it will cause a NullPointerException in the
engine; breaking forward compatibility.

This patch populates the imageID property to the disk metadata for LUNs
disks, with the LUN's GUID. It will resolve the forward compatibility to
the engine.

Change-Id: Iaf654416990afe7c9e0a9cb8d8b99231b80a0e77
Bug-Url: https://bugzilla.redhat.com/1859092
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>