storage.misc.walk: name skipped paths appropriately
misc.walk() is asked to os.walk(), but never get into particular paths
that the caller would like to avoid. These unwanted paths do not have a
particularly low albedo value; they are not really black. And they do
not necessarily come as a list. So let us name the argument for what it
is: a bunch of stuff we want to skip.

Change-Id: I8cea5a3e3456ae2a85fe3e54732c174a81b8c669
Signed-off-by: Dan Kenigsberg <danken@redhat.com>