sd: Change validateCreateVolumeParams to method
StorageDomainManifest.validateCreateVolumeParams() was changed to a
class method in:

commit a28018598ac408b741b9f09ace3adea09c132d2a
Author: Denis Chaplygin <dchaplyg@redhat.com>
Date:   Mon Jan 28 16:54:24 2019 +0100

    storage: Updated DiskType in VolumeMetadata

To make it easier to test. This was fine at that time, but now we want
to check also the storage domain version. This requires passing the
domain version via multiple layers. Convert the method back to instance
method to make this easier.

Change-Id: Id6fa5f1de0e15893019d718e787c2ca8e09f82a5
Signed-off-by: Nir Soffer <nsoffer@redhat.com>